# server/routes/api/docs/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Scriptor Service
  license:
    name: Copyright 2015-20
tags:
  - name: Locator
    description: 'Activities related to single step template and script creation'
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
host: sim5runner.herokuapp.com
basePath: /api
paths:
  '/locators':
    post:
      tags:
        - Locator
      summary: Post in a group
      operationId: postLocator
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: 'Post object to be added to the group(userid, email and name is optional)'
          required: true
          schema:
            $ref: '#/definitions/locator'
      responses:
        '200':
          description: Post added to group
          schema:
            type: array
            items:
              $ref: '#/definitions/locator'
        '400':
          description: Bad request
        '401':
          description: Invalid Authorization Token
        '403':
          description: User not authorized
        '404':
          description: Resource not found
  '/locators/{app}':
    get:
      tags:
        - Locator
      summary: Get locators
      operationId: getLocators
      description: ''
      produces:
        - application/json
      parameters:
        - name: app
          in: path
          description: App type of locator
          required: true
          type: string
      responses:
        '200':
          description: locators
          schema:
            $ref: '#/definitions/locators'
        '400':
          description: Bad request
        '401':
          description: Invalid Authorization Token
        '403':
          description: User not authorized
        '404':
          description: Resource not found
  '/locators/{app}/{key}':  
    get:
      tags:
        - Locator
      summary: Get locator by key
      operationId: getLocatorByKey
      description: ''
      produces:
        - application/json
      parameters:
        - name: app
          in: path
          description: App type of locator
          required: true
          type: string
        - name: key
          in: path
          description: App type of locator
          required: true
          type: string
      responses:
        '200':
          description: locators
          schema:
            $ref: '#/definitions/locator'
        '400':
          description: Bad request
        '401':
          description: Invalid Authorization Token
        '403':
          description: User not authorized
        '404':
          description: Resource not found



definitions:
  locators:
    required:
      - locators
    properties:
      locators:
        type: array
        items:
          $ref: '#/definitions/locator'
  locator:
    required:
      - app
      - tags
      - xpath
    properties:
      app:
        type: string
      tags:
        type: string
      xpath:
        type: object
        $ref: '#/definitions/xpathSchema'
  xpathSchema:
    required:
      - key
      - value
    properties:
      key:
        type: string
      value:
        type: string